% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingdata.R
\name{loadingdata}
\alias{loadingdata}
\title{Load and prepare data for analysis.}
\usage{
loadingdata(
  fileClinicalDataPath,
  folderImagingDataPath,
  indVar,
  nPoints,
  organ
)
}
\arguments{
\item{fileClinicalDataPath}{is the path of the file containing the clinical data of the subjects for all the models.}

\item{folderImagingDataPath}{is the path of the folders storing the image data e.g. "/home/yourname/folder-path/".}

\item{indVar}{is the independent variables of the model e.g. c("individual_id","Age","Sex","Race_c","BMI","WHR").}

\item{nPoints}{is the number of points (vertices) in the mesh, this will change for each organ e.g. for kidney left: 4380.}

\item{organ}{is the organ segmentation e.g. liver, spleen, kidney_left etc.}
}
\description{
Load and prepare data for analysis.
}
\examples{
model <- loadingData(fileClinicalDataPath, folderImagingDataPath, indVar, nPoints, organ)
}
