% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_univariate_analysis.R
\name{mass_univariate_analysis}
\alias{mass_univariate_analysis}
\title{Mass Univariate Analysis.}
\usage{
mass_univariate_analysis(
  data_dir,
  phenotype,
  organ,
  scale_range,
  nPermutations,
  data_range,
  extract_range,
  output_dir
)
}
\arguments{
\item{data_dir}{is the path of the data directory were the matrices X and Y for the model are saved.}

\item{phenotype}{is the imaging phenotype e.g. 1 for S2S, 2 for Curvature.}

\item{organ}{is the organ segmentation e.g. liver, spleen, kidney_left etc.}

\item{scale_range}{is the range of the independent variables to be scaled e.g. c(1:10).}

\item{nPermutations}{is the number of permutation testing e.g. 1000.}

\item{data_range}{is the range of the independent variables to be included in the model e.g. c(1:8).}

\item{extract_range}{is the range of the independent variables to extract the beta coefficients and p-values after MUR analysis e.g. c(2, 5:8).}

\item{output_dir}{is the path of the output directory were the output with the coefficients and p-values for each covariate are saved as a ```.txt``` file.}
}
\description{
Mass Univariate Analysis.
}
\examples{
mur_analysis <- mass_univariate_analysis(data_dir, phenotype, organ, scale_range, nPermutations, data_range, extract_range, output_dir)
}
