% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permFL_fast.R
\name{permFL_fast}
\alias{permFL_fast}
\title{Compute TFCE derived p-values using Freedman-Lane procedure faster!}
\usage{
permFL_fast(X, Y, extract, A, NNmatrix, nPermutations, E = 0.5, H = 2)
}
\arguments{
\item{X}{is the design matrix. Number of rows = number of subjects in the study, number of columns = number of vertices in the atlas. Numerical varable must be normalized to 0-mean and unit-standard deviation. Categorical variables must be coded using dummy coding. The first column should contain the intercept (all 1s).}

\item{Y}{is the imaging matrix. Number of rows = N. Number of columns = V.}

\item{extract}{is an array expressing which covariates in X you want to extract.}

\item{A}{A V-dimensional vector containing the area associated with a vertex, usually its Voronoi area.}

\item{NNmatrix}{Nx2 matrix containing the mesh edges. Important: to speed up the execution please avoid repetitions like (A,B) and (B,A).}

\item{nPermutations}{number of permutations in the permutation test, default is 1000.}

\item{E}{is the TFCE parameter, by default fixed to 0.5.}

\item{H}{is the TFCE parameter, by default fixed to 2.}
}
\description{
Given a NxV imaging matrix Y (N = number of subjects, V = number of vertices in the ventricular mesh), a NxC model matrix X  (N = number of subjects, C = number of variables + intercept term)
and the numbers of the column variables to extract, this function computes for each variable specified in extract a couple of arrays: in the first array the TFCE derived p-values map on the mesh and in the second array a binary variable expressing whether that vertex has reached significance using a whole-image threshold as specified in the TFCE original paper. The output is a matrix with a number of
columns equal to twice the lenght of extract and a number of rows equal to the number of vertices.
}
\examples{
TFCEresults = permFL_fast(X, Y, extract, A, NNmatrix, nPermutations = 1000, E=0.5, H=2)
}
